chore: rename algolia hook
