+++ src/collections/CommunityHelp/index.ts
@@ -2,7 +2,7 @@ import type { CollectionConfig } from 'payload/types'
 
 import { isAdmin } from '../../access/isAdmin'
 import { extractDescription } from './extract-description'
-import { removeFromAlgolia } from './removeFromAlgolia'
+import { updateAlgolia } from './updateAlgolia'
 
 export const CommunityHelp: CollectionConfig = {
   slug: 'community-help',
@@ -108,7 +108,7 @@ export const CommunityHelp: CollectionConfig = {
                 siblingData.communityHelpType === 'discord'
                   ? siblingData.discordID
                   : siblingData.githubID
-              removeFromAlgolia(docID, value)
+              updateAlgolia(docID, value)
             }
           },
         ],
