+++ src/collections/CommunityHelp/updateAlgolia.ts
